requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
llama-index==0.10.28
python-dotenv==1.0.1
traceloop-sdk==0.16.1
gunicorn==22.0.0